python-send-push/pyproject.toml
[tool.poetry]
name = "firebase-push"
version = "0.1.0"
description = ""
authors = ["Artur Akmalov <artur@akmalov.com>"]
readme = "README.md"
packages = [{include = "firebase_push"}]

[tool.poetry.dependencies]
python = "^3.11"
google-auth = "^2.30.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
